requirements.txt
Flask==3.1.3
rawpy==0.26.1
Pillow==11.3.0
pdf2docx==0.5.12
moviepy==2.2.1
werkzeug==3.1.8
requests==2.33.1
yt-dlp==2026.3.17
gunicorn==21.2.0
